+++ TopTagger/include/TTMHEPRequirements.h
@@ -8,7 +8,7 @@ class TopTaggerResults;
 class TTMHEPRequirements : public TTModule
 {
 private:
-    double mW_, mt_, Rmin_, Rmax_, csvThresh_, bEtaCut_;
+    float mW_, mt_, Rmin_, Rmax_, csvThresh_, bEtaCut_;
     int maxNbInTop_;
     bool doTrijet_, doDijet_, doMonojet_;
