+++ TopTagger/include/TopObject.h
@@ -13,7 +13,7 @@ class TopObject
 {
 private:
     TLorentzVector p_;
-    double dRmax_, discriminator_;
+    float dRmax_, discriminator_;
     std::vector<Constituent const *> constituents_;
 
     void updateVariables();
@@ -25,19 +25,19 @@ public:
     TopObject(std::vector<Constituent const *> constituents);
     
     void addConstituent(Constituent const *  constituent);
-    void setDiscriminator(const double disc) { discriminator_ = disc; }
+    void setDiscriminator(const float disc) { discriminator_ = disc; }
 
     const TLorentzVector& p() const { return p_; }
     const TLorentzVector& P() const { return p(); }
-    double getDRmax() const { return dRmax_; }
-    double getDiscriminator() const { return discriminator_; }
+    float getDRmax() const { return dRmax_; }
+    float getDiscriminator() const { return discriminator_; }
 
     const std::vector<Constituent const *>& getConstituents() const { return constituents_; }
     int getNConstituents() const { return constituents_.size(); }
-    int getNBConstituents(double cvsCut, double etaCut = 2.4) const;
+    int getNBConstituents(float cvsCut, float etaCut = 2.4) const;
 
     const decltype(genMatchPossibilities_)& getGenTopMatches() const { return genMatchPossibilities_; }
-    const TLorentzVector* getBestGenTopMatch(const double dRMax = 0.6) const;
+    const TLorentzVector* getBestGenTopMatch(const float dRMax = 0.6) const;
 };
 
 #endif
