+++ TopTagger/src/TTMBasicClusterAlgo.cpp
@@ -62,7 +62,7 @@ void TTMBasicClusterAlgo::run(TopTaggerResults& ttResults)
                     TopObject topCand({&constituents[i], &constituents[j]});
 
                     //mass window on the top candidate mass
-                    double m123 = topCand.p().M();
+                    float m123 = topCand.p().M();
                     bool passMassWindow = (minTopCandMass_ < m123) && (m123 < maxTopCandMass_);
 
                     if(topCand.getDRmax() < dRMaxDiJet_ && passMassWindow)
@@ -89,7 +89,7 @@ void TTMBasicClusterAlgo::run(TopTaggerResults& ttResults)
                                 TopObject topCand({&constituents[k], &constituents[j], &constituents[i]});
 
                                 //mass window on the top candidate mass
-                                double m123 = topCand.p().M();
+                                float m123 = topCand.p().M();
                                 bool passMassWindow = (minTopCandMass_ < m123) && (m123 < maxTopCandMass_);
 
                                 if(topCand.getDRmax() < dRMaxTrijet_ && passMassWindow)
